api/pyproject.toml
[tool.poetry]
name = "llm-domain-framework-api"
version = "0.1.0"
description = "LLM Domain Framework API"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
aiohttp = "^3.9.0"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
numpy = "^1.24.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
alembic = "^1.12.0"
greenlet = "^3.0.0"

aiofiles = "^24.1.0"
python-multipart = "^0.0.20"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"

pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
download-models = "scripts.download_models:download_models"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::sqlalchemy.exc.MovedIn20Warning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
] 